# specs/bff-openapi.yaml
openapi: 3.1.0
info:
  title: Fantasy Hockey BFF API
  description: Backend for Frontend aggregation layer for the Fantasy Hockey application
  version: "1.0"
servers:
- url: /
security:
- bearerAuth: []
...
paths:
  /api/v1/projections/{id}:
    get:
      tags:
      - Projections
      summary: "Fetch one of the current user's saved projections, including its data"
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Projection found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "404":
          description: No such projection for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
    put:
      tags:
      - Projections
      summary: Update one of the current user's saved projections
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectionRequest"
        required: true
      responses:
        "200":
          description: Projection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "404":
          description: No such projection for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "409":
          description: Another projection with that name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
    delete:
      tags:
      - Projections
      summary: Delete one of the current user's saved projections
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Projection deleted
        "404":
          description: No such projection for this user
  /api/v1/projections/{id}/share:
    get:
      tags:
      - Projection shares
      summary: Fetch the share link for one of the current user's projections
      operationId: getProjectionShare
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: The projection is shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
        "404":
          description: "No such projection for this user, or it is not shared"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
    put:
      tags:
      - Projection shares
      summary: Publish or refresh the projection's public snapshot
      description: "Idempotent: re-sharing keeps the same link and updates what it\
        \ shows."
      operationId: shareProjection
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareProjectionRequest"
        required: true
      responses:
        "200":
          description: Share created or refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
        "404":
          description: No such projection for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
    delete:
      tags:
      - Projection shares
      summary: Take the projection's public link down
      description: "The link stops working immediately, and sharing again produces\
        \ a new one."
      operationId: unshareProjection
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Share deleted
        "404":
          description: "No such projection for this user, or it is not shared"
...
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh tokens
      description: Exchange a valid refresh token for a new access token and refresh
        token pair
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/shared/{token}:
    get:
      tags:
      - Shared projections
      summary: Fetch a shared projection by its token
      description: "Public: anyone holding the link can read it. Returns the snapshot\
        \ as it was when shared, with no identity beyond the alias the owner chose."
      operationId: getSharedProjection
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Shared projection found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedProjectionResponse"
        "404":
          description: "No share with that token, or it was taken down"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedProjectionResponse"
...
components:
  schemas:
    DraftPick:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
        teamId:
          type: string
          maxLength: 64
          minLength: 0
      required:
      - playerId
      - teamId
    DraftState:
      type: object
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/DraftTeam"
          maxItems: 32
          minItems: 0
        order:
          type: array
          items:
            type: string
            maxLength: 64
            minLength: 0
          maxItems: 32
          minItems: 0
        picks:
          type: array
          items:
            $ref: "#/components/schemas/DraftPick"
        finishedAt:
          type: string
          format: date-time
      required:
      - order
      - picks
      - teams
    DraftTeam:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
          minLength: 0
        name:
          type: string
          maxLength: 64
          minLength: 0
        mine:
          type: boolean
      required:
      - id
      - mine
      - name
    PlayerProjection:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
        type:
          type: string
          enum:
          - skater
          - goalie
        stats:
          $ref: "#/components/schemas/PlayerStats"
      required:
      - playerId
      - stats
      - type
    PlayerStats:
      type: object
      properties:
        utility:
          type: object
          additionalProperties:
            type: number
            format: double
        scoring:
          type: object
          additionalProperties:
            type: number
            format: double
      required:
      - scoring
      - utility
    ProjectionSettings:
      type: object
      properties:
        scoringType:
          type: string
          enum:
          - points
          - category
        statWeights:
          type: object
          additionalProperties:
            type: number
            format: double
        activeScoringColumns:
          type: array
          items:
            type: string
        activeUtilityColumns:
          type: array
          items:
            type: string
        scaleSettings:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ScaleConfig"
        decimalSettings:
          type: object
          additionalProperties:
            type: integer
            format: int32
        useDefaultDecimals:
          type: boolean
        leagueSize:
          type: integer
          format: int32
          maximum: 30
          minimum: 2
        rosterSlots:
          $ref: "#/components/schemas/RosterSlots"
        minGoalieGames:
          type: integer
          format: int32
          maximum: 82
          minimum: 0
        yahooSync:
          $ref: "#/components/schemas/YahooSync"
      required:
      - activeScoringColumns
      - activeUtilityColumns
      - decimalSettings
      - scaleSettings
      - scoringType
      - statWeights
      - useDefaultDecimals
    RosterSlots:
      type: object
      properties:
        c:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        lw:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        rw:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        d:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        util:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        bn:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        g:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
      required:
      - bn
      - c
      - d
      - g
      - lw
      - rw
      - util
    ScaleConfig:
      type: object
      properties:
        scale:
          type: boolean
        scalableStats:
          type: array
          items:
            type: string
      required:
      - scalableStats
      - scale
    UpdateProjectionData:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/ProjectionSettings"
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerProjection"
        draft:
          $ref: "#/components/schemas/DraftState"
        tings:
          $ref: "#/components/schemas/UpdateProjectionData"
          writeOnly: true
      required:
      - settings
    UpdateProjectionRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        data:
          $ref: "#/components/schemas/UpdateProjectionData"
      required:
      - data
      - name
    YahooSync:
      type: object
      properties:
        leagueName:
          type: string
          maxLength: 200
          minLength: 0
        leagueKey:
          type: string
          maxLength: 100
          minLength: 0
        syncedAt:
          type: string
          format: date-time
      required:
      - leagueKey
      - leagueName
      - syncedAt
    ProjectionData:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/ProjectionSettings"
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerProjection"
        draft:
          $ref: "#/components/schemas/DraftState"
        tings:
          $ref: "#/components/schemas/ProjectionData"
          writeOnly: true
      required:
      - players
      - settings
    ProjectionResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum:
          - projection
          - preset_draft
        season:
          type: string
          enum:
          - "20252026"
          - "20262027"
          - "20272028"
          - "20282029"
        data:
          $ref: "#/components/schemas/ProjectionData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - data
      - id
      - kind
      - name
      - season
      - updatedAt
    ShareProjectionRequest:
      type: object
      properties:
        authorAlias:
          type: string
          description: Name to be credited as on the public page. Omit to be credited
            as nobody — the account's email is never shown.
          maxLength: 40
          minLength: 0
        players:
          type: array
          description: "The ranked rows to publish, in display order."
          items:
            $ref: "#/components/schemas/SharedPlayer"
          maxItems: 200
          minItems: 0
      required:
      - players
    SharedPlayer:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 100
          minLength: 0
        teamAbbrev:
          type: string
          maxLength: 10
          minLength: 0
        positions:
          type: array
          items:
            type: string
            maxLength: 4
            minLength: 0
          maxItems: 6
          minItems: 0
        type:
          type: string
          enum:
          - skater
          - goalie
        rank:
          type: integer
          format: int32
          minimum: 1
        value:
          type: number
          format: double
        stats:
          $ref: "#/components/schemas/PlayerStats"
      required:
      - name
      - playerId
      - rank
      - stats
      - type
      - value
    ShareLinkResponse:
      type: object
      properties:
        token:
          type: string
        shareUrl:
          type: string
          description: "The public link to share, e.g. https://slapstat.com/s/<token>"
        authorAlias:
          type: string
          description: "The name the public page credits, if the owner set one."
        viewCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - shareUrl
      - token
      - updatedAt
      - viewCount
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
        refreshToken:
          type: string
        refreshExpiresInSeconds:
          type: integer
          format: int64
        admin:
          type: boolean
        emailVerified:
          type: boolean
          description: Whether the account's email address has been verified
      required:
      - admin
      - emailVerified
      - expiresInSeconds
      - refreshExpiresInSeconds
      - refreshToken
      - token
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          maxLength: 4096
          minLength: 0
      required:
      - refreshToken
...
    SharedProjectionData:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/ProjectionSettings"
        players:
          type: array
          items:
            $ref: "#/components/schemas/SharedPlayer"
        tings:
          $ref: "#/components/schemas/SharedProjectionData"
          writeOnly: true
      required:
      - players
      - settings
    SharedProjectionResponse:
      type: object
      properties:
        token:
          type: string
        name:
          type: string
        authorAlias:
          type: string
        season:
          type: string
          enum:
          - "20252026"
          - "20262027"
          - "20272028"
          - "20282029"
        data:
          $ref: "#/components/schemas/SharedProjectionData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - data
      - name
      - season
      - token
      - updatedAt
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT